requirements.txt
google_api_python_client==2.131.0
google_auth_oauthlib==1.2.0
langchain==0.2.2
nicegui==1.4.26
openai==1.31.1
protobuf==3.19.1
python-dotenv==1.0.1
